Added store_id as parameter
As core_url_rewrite stores rewrites based on store_id and request_path
we need both to identify a correct target. Added lookup of default
store_id=0 to the request to handle unspecific redirects at all. Order
by is needed to ensure, if default and store specific rewrites are found,
the default one is skipped and the more specific is used.